Iterate the input only
Before, we had the instructions we were iterating.
We also had the stack we were iterating.
Instead, we pass the stack around as the state,
and iterate only the instructions.

Part of the problem with the previous implementation was that we would
have to restart iteration of the instructions from the beginning if
we ran into a user-land word.

We also fix the naming to more accurately reflect what each
`[]string` represents.