Handle bounds for a definition without a word
We weren't checking that we had enough input to consume the next word
as the name of the word.
I.e. There was a `:` as the last word in the input line.

To address this, we use a magic string that cannot be a valid definiton.
Then we case on that string to see if we need to consume the first word
of the input as the name of the word.